refactor commands to command_modules and actions
