+++ src/cli/actions/index.ts
@@ -1,0 +1,5 @@
+export * from "./check";
+export * from "./copy";
+export * from "./listStates";
+export * from "./listCommands";
+export * from "./snip";
