+++ src/cli/actions/listStates.ts
@@ -1,10 +1,9 @@
-import { Arguments, CommandModule } from "yargs";
-import { getBluehawk } from "../../../bluehawk";
-import { MainArgs } from "../../cli";
-import { withIgnoreOption, withJsonOption } from "../../options";
-import { printJsonResult } from "../../printJsonResult";
+import { Arguments } from "yargs";
+import { getBluehawk } from "../../bluehawk";
+import { MainArgs } from "../cli";
+import { printJsonResult } from "../printJsonResult";
 
-interface ListStatesArgs extends MainArgs {
+export interface ListStatesArgs extends MainArgs {
   paths: string[];
   json?: boolean;
   ignore?: string | string[];
@@ -47,20 +46,3 @@ export const listStates = async (
       .join("\n")}`
   );
 };
-
-const commandModule: CommandModule<
-  MainArgs & { paths: string[] },
-  ListStatesArgs
-> = {
-  command: "states <paths..>",
-  builder(argv) {
-    return withJsonOption(withIgnoreOption(argv));
-  },
-  async handler(args) {
-    return await listStates(args);
-  },
-  aliases: [],
-  describe: "list states used in the given project",
-};
-
-export default commandModule;
