+++ src/cli/actions/snip.test.ts
