+++ src/cli/command_modules/check.ts
