+++ src/cli/command_modules/index.ts
@@ -1,5 +1,6 @@
 export * from "./check";
 export * from "./copy";
-export * from "./list/";
 export * from "./list";
+export * from "./listCommands";
+export * from "./listStates";
 export * from "./snip";
