--- src/cli/commands/list/index.ts
@@ -1,2 +1,0 @@
-export * from "./commands";
-export * from "./states";
