+++ src/cli/index.ts
@@ -1,5 +1,5 @@
 export * from "./Plugin";
 export * from "./cli";
-export * from "./commands";
+export * from "./actions";
 export * from "./options";
 export * from "./printJsonResult";
